reformatted

+++ server/actions.py
@@ -1,9 +1,10 @@
-'''
+"""
 actions.py: Main logic of actions to process after recieved packages.
-'''
+"""
 from common.utils import logger
 from sys import exit
 import json
+import os
 from server.config import *
 
 # resolve SQL connection
@@ -14,7 +15,7 @@ elif SQL_TYPE == SQLTYPE.MARIADB:
 elif SQL_TYPE == SQLTYPE.NONE:
 
     def connect(**kwargs):
-        logger.error('No SQL selected. Cannot publish any connection.')
+        logger.error("No SQL selected. Cannot publish any connection.")
         exit(-1)
 
     class Error(Exception):
@@ -22,31 +23,33 @@ elif SQL_TYPE == SQLTYPE.NONE:
 
 
 def change_question_mark(sentence):
-    return sentence[:-1] + '!'
+    return sentence[:-1] + "!"
 
 
 def database_test(sql):
     try:
-        conn = connect(user=SQL_USER,
-                       password=SQL_PASSWORD,
-                       host=SQL_ADDRESS,
-                       port=SQL_PORT)
+        conn = connect(
+            user=SQL_USER, password=SQL_PASSWORD, host=SQL_ADDRESS, port=SQL_PORT
+        )
     except Error as e:
-        logger.error(
-            f'Error connecting to {SQL_TYPE.name.lower().title()}: {e}')
-        return ''
-    logger.info(f'Successfully connected to {SQL_TYPE.name.lower().title()}.')
+        logger.error(f"Error connecting to {SQL_TYPE.name.lower().title()}: {e}")
+        return ""
+    logger.info(f"Successfully connected to {SQL_TYPE.name.lower().title()}.")
     cur = conn.cursor()
     cur.execute(sql)
-    res = ''
+    res = ""
     for d in cur:
-        res += d[0] + '\n'
+        res += d[0] + "\n"
     return res[:-1]
+
 
 # write your own actions here
 
+
 def login(username, password):
-    with open('users.json', 'r') as f:
+    if not os.path.exists("users.json"):
+        return False
+    with open("users.json", "r") as f:
         users = json.load(f)
     if username in users and users[username] == password:
         return True
@@ -55,12 +58,16 @@ def login(username, password):
 
 
 def register(username, password):
-    with open('users.json', 'r') as f:
+    if not os.path.exists("users.json"):
+        with open("users.json", "w") as f:
+            json.dump({username: password}, f)
+        return True
+    with open("users.json", "r") as f:
         users = json.load(f)
     if username in users:
         return False
     else:
         users[username] = password
-        with open('users.json', 'w') as f:
+        with open("users.json", "w") as f:
             json.dump(users, f)
         return True
